Added three types of launch
